Add SSE support
Add an Server-Sent Event handler, utilizing the EventStreams
(stream.wikimedia.org) service, and the aiohttp-sse-client Python
library.

This allows ircstream to be used outside of the MediaWiki Wikimedia
production infrastructure, or any MediaWiki infrastructure for that
matter. Events will be consumed in realtime using HTTP SSE, formatted
according to the long-standing legacy format, and broadcasted to
connected IRC clients.

This is considered WIP, as there are several limitations:
* stream.wikimedia.org has relatively short timeouts and disconnects
  clients periodically. While resumption
  closed is possible, it's not currently possible to do so due to both
  server and client library limitations. Therefore, messages may (and
  most likely will) be lost between disconnection and reconnection.
* A lot of effort has gone into adequate error handling in general but
  the code has not been battle-tested. Expect (but report) crashes.
* There are other missing features such as Special page localization and
  #centralauth that will be different from rc2udp.