sse: replace aiohttp_sse_client with a custom implementation
The aiohttp_sse_client library was a bit of a pain to use, as were the
other libraries it was inspired from (aiosseclient, sseclient, and
sseclient-py).

(I know because I've used *all of them* over the years, through commits
that never saw the day. This is implementation v5).

The code is loosely based on aiohttp_sse_client. Overall, this:
* Simplifies a lot of the error handling by avoiding
  try/except/raise/try/except chains; exceptions are handled closer to
  the caller now.
* Makes the code easier to reason about, e.g. EventSource was both an
  async context manager and an async iterator.
* Modernizes the code a bit (attrs -> dataclasses)
* Avoids the extra subclassing/overrides, and makes the last_event_id
  persistence much nicer.
* Reduces the total SLOC count, and hopefully number of bugs, by quite a
  bit.