Generalize IRCClient's background tasks functionality
Right now we only have one background task: the periodic ping. In
anticipation of adding another one, generalize the existing code to keep
the tasks in a set rather than a single variable, and to cancel all
members of that set on shutdown.

Incidentally, this makes the code more similar to the Python
documentation's reference code, as well as main.py.