Workaround PING bug with SmartIrc4net (CVNBot)
Per modern IRC specs:
> Parameters that contain spaces, are empty, or begin with a ':'
> character MUST be sent with a preceding ':'; in other cases the use of a
> preceding ':' on the final parameter is OPTIONAL.

The BNF in RFCs 1459 and 2812 also match that description.

Many servers seem to send the (optional) preceding colons semi-randomly.
For instance, ircd-ratbox-2.2.9 sends preceding colons for JOINs, but
not PARTs.

Similarly, some client libraries do not seem to be consistent to this
regard as well. meebey/SmartIrc4net (used by e.g. CVNBot) in particular
is not very smart *g*, and seems to have implemented message parsing using
regexps(!). The regexps were seemingly  built based on observations of
the wire protocol rather than implementing the specification.
Therefore, its source code includes this gem:
  private static Regex     _PingRegex               = new Regex("^PING :.*", RegexOptions.Compiled);

ircstream has a robust standards-compliant parser and emitter, tested
rigorously with the IRC parser tests. As such, it emits the (perfectly
valid) "PING servername" on PINGs, without a colon. That string,
however, is never matched by SmartIrc4net's regexp.

Introduce a hack here to always send colons with PINGs in particular, to
ensure byte-for-byte compatibility with other servers that tend to emit
that, and client libraries used in the wild.