Stop recommending EventStreams
EventStreams, including but not limited to the recentchanges stream, has
been poorly maintained, and is riddled with bugs, missing features, and
architectural limitations. No official SDK exists, only a sprawling
number of third-party authored, semi-maintained libraries, many with
unanswered GitHub issues of users having random encounters with some of
the issues that I've stumbled upon myself.

Most of these issues were reported as early as 2016, when the original
ircstream was developed, and the vast majority of which have remained
unfixed for 8+ years now. The few that have been fixed were drive-by
contributions, done mostly as a favor to me.

Today the situation has actually worsened: at least one more data bug
has crept in, data loss of >1% has been observed compared to RC2UDP, and
the service's ownership status is (to the shock of noone) in an even
greater limbo. Multiple overlapping streams, offering similar
functionality, but none offering feature parity with RC2UDP, exist, with
a varying degree of exposure and support to third-party users.

I've had to reimplement the SSE client more times than I can count, and
ended up with an implementation that kinda works but has a number of
hacks to workaround all the random undocumented SSE limitations. It's
still severely limited, and cannot offer feature parity with RC2UDP. No
downstream user should ever have to go through that.

It's therefore both disingenuous and misleading to suggest to ircstream
users to use EventStreams, so stop doing so.